+++ lib/lightning/projects/project_user.ex
@@ -23,6 +23,13 @@ defmodule Lightning.Projects.ProjectUser do
     :owner
   ])
 
+  defenum(DigestEnum, :digest, [
+    :never,
+    :daily,
+    :weekly,
+    :monthly
+  ])
+
   @primary_key {:id, :binary_id, autogenerate: true}
   @foreign_key_type :binary_id
   schema "project_users" do
@@ -31,6 +38,7 @@ defmodule Lightning.Projects.ProjectUser do
     field :delete, :boolean, virtual: true
     field :failure_alert, :boolean, default: true
     field :role, RolesEnum, default: :editor
+    field :digest, DigestEnum, default: :weekly
 
     timestamps()
   end
@@ -42,7 +50,7 @@ defmodule Lightning.Projects.ProjectUser do
   @doc false
   def changeset(project_user, attrs) do
     project_user
-    |> cast(attrs, [:user_id, :project_id, :role, :failure_alert])
+    |> cast(attrs, [:user_id, :project_id, :role, :digest, :failure_alert])
     |> validate_required([:user_id])
     |> unique_constraint([:project_id, :user_id],
       message: "User already a member of this project."