+++ lib/lightning_web/live/project_live/settings.ex
@@ -4,6 +4,10 @@ defmodule LightningWeb.ProjectLive.Settings do
   """
   use LightningWeb, :live_view
 
+  alias Lightning.Policies.Permissions
+  alias Lightning.Accounts.User
+  alias Lightning.Policies.ProjectUsers
+  alias Lightning.Projects.ProjectUser
   alias Lightning.{Projects, Credentials}
 
   on_mount({LightningWeb.Hooks, :project_scope})
@@ -33,8 +37,17 @@ defmodule LightningWeb.ProjectLive.Settings do
        can_edit_project: can_edit_project,
        credentials: credentials,
        project_users: project_users,
-       changeset: Projects.change_project(socket.assigns.project)
+       current_user: socket.assigns.current_user,
+       project_changeset: Projects.change_project(socket.assigns.project)
      )}
+  end
+
+  def can_edit_project_user(
+        %User{} = current_user,
+        %ProjectUser{} = project_user
+      ) do
+    ProjectUsers
+    |> Permissions.can(:edit_project_user, current_user, project_user)
   end
 
   @impl true
@@ -53,11 +66,141 @@ defmodule LightningWeb.ProjectLive.Settings do
       |> Projects.change_project(project_params)
       |> Map.put(:action, :validate)
 
-    {:noreply, assign(socket, :changeset, changeset)}
+    {:noreply, assign(socket, :project_changeset, changeset)}
   end
 
   def handle_event("save", %{"project" => project_params} = _params, socket) do
     save_project(socket, project_params)
+  end
+
+  def handle_event(
+        "set_failure_alert",
+        %{
+          "project_user_id" => project_user_id,
+          "value" => value
+        } = _params,
+        socket
+      ) do
+    project_user = Projects.get_project_user!(project_user_id)
+
+    changeset =
+      {%{failure_alert: project_user.failure_alert}, %{failure_alert: :boolean}}
+      |> Ecto.Changeset.cast(%{failure_alert: value}, [:failure_alert])
+
+    case Ecto.Changeset.get_change(changeset, :failure_alert) do
+      nil ->
+        {:noreply, socket}
+
+      setting ->
+        Projects.update_project_user(project_user, %{failure_alert: setting})
+        |> dispatch_flash(socket)
+    end
+  end
+
+  def handle_event(
+        "set_digest",
+        %{"project_user_id" => project_user_id, "value" => value},
+        socket
+      ) do
+    project_user = Projects.get_project_user!(project_user_id)
+
+    changeset =
+      {%{digest: project_user.digest |> Atom.to_string()}, %{digest: :string}}
+      |> Ecto.Changeset.cast(%{digest: value}, [:digest])
+
+    case Ecto.Changeset.get_change(changeset, :digest) do
+      nil ->
+        {:noreply, socket}
+
+      digest ->
+        Projects.update_project_user(project_user, %{digest: digest})
+        |> dispatch_flash(socket)
+    end
+  end
+
+  defp dispatch_flash(change_result, socket) do
+    case change_result do
+      {:ok, %ProjectUser{}} ->
+        {:noreply,
+         socket
+         |> assign(
+           :project_users,
+           Projects.get_project_with_users!(socket.assigns.project.id).project_users
+         )
+         |> put_flash(:info, "Project user updated successfuly")}
+
+      {:error, %Ecto.Changeset{}} ->
+        {:noreply,
+         socket
+         |> put_flash(:error, "Error when updating the project user")}
+    end
+  end
+
+  def failure_alert(assigns) do
+    ~H"""
+    <%= if can_edit_project_user(@current_user, @project_user) do %>
+      <select
+        id={"failure-alert-#{@project_user.id}"}
+        phx-change="set_failure_alert"
+        phx-value-project_user_id={@project_user.id}
+        class="mt-1 block w-full rounded-md border-secondary-300 shadow-sm text-sm focus:border-primary-300 focus:ring focus:ring-primary-200 focus:ring-opacity-50"
+      >
+        <%= options_for_select(
+          [Disabled: "false", Enabled: "true"],
+          @project_user.failure_alert
+        ) %>
+      </select>
+    <% else %>
+      <%= if @project_user.failure_alert,
+        do: "Enabled",
+        else: "Disabled" %>
+    <% end %>
+    """
+  end
+
+  def digest(assigns) do
+    # you will get a form
+    # assigns.form.source.project_user
+    assigns =
+      assigns
+      |> assign(
+        can_edit_project_user:
+          can_edit_project_user(assigns.current_user, assigns.project_user)
+      )
+
+    ~H"""
+    <%= if @can_edit_project_user do %>
+      <select
+        id={"digest-#{@project_user.id}"}
+        phx-change="set_digest"
+        phx-value-project_user_id={@project_user.id}
+        class="mt-1 block w-full rounded-md border-secondary-300 shadow-sm text-sm focus:border-primary-300 focus:ring focus:ring-primary-200 focus:ring-opacity-50"
+      >
+        <%= options_for_select(
+          [Never: "never", Daily: "daily", Weekly: "weekly", Monthly: "monthly"],
+          @project_user.digest
+        ) %>
+      </select>
+    <% else %>
+      <%= @project_user.digest
+      |> Atom.to_string()
+      |> String.capitalize() %>
+    <% end %>
+    """
+  end
+
+  def role(assigns) do
+    ~H"""
+    <%= @project_user.role
+    |> Atom.to_string()
+    |> String.capitalize() %>
+    """
+  end
+
+  def user(assigns) do
+    ~H"""
+    <%= @project_user.user.first_name %> <%= @project_user.user.last_name %>
+    """
   end
 
   defp save_project(socket, project_params) do
@@ -69,7 +212,7 @@ defmodule LightningWeb.ProjectLive.Settings do
          |> put_flash(:info, "Project updated successfully")}
 
       {:error, %Ecto.Changeset{} = changeset} ->
-        {:noreply, assign(socket, :changeset, changeset)}
+        {:noreply, assign(socket, :project_changeset, changeset)}
     end
   end
 end